backend/requirements.txt
# Core
fastapi==0.110.1
uvicorn==0.25.0
gunicorn>=21.2.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic>=2.6.4
email-validator>=2.2.0

# Database
pymongo==4.5.0
motor==3.3.1

# Auth
pyjwt>=2.10.1
passlib[bcrypt]>=1.7.4
cryptography>=42.0.8

# HTTP / Scraping
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.13.0

# PDF processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# Scheduling
apscheduler>=3.10.4

# Deduplication
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0

# Notifications
aiosmtplib>=2.0.0

# Timezone
tzdata>=2024.2
